Public API

Contents

Public API

Contents

Index

Public interface

Index

EquationOfStateRecipes.bulkmodulusplot

EquationOfStateRecipes.energyplot

EquationOfStateRecipes.energypressureplot

EquationOfStateRecipes.pressureplot

RecipesBase.plot

Public interface

For detailed examples, please refer to section "Examples".

RecipesBase.plot — Function

plot(eos::EnergyEquation, volumes, args...; kw...)
plot(eos::PressureEquation, volumes, args...; kw...)
plot(eos::BulkModulusEquation, volumes, args...; kw...)
plot!(eos::EnergyEquation, volumes, args...; kw...)
plot!(eos::PressureEquation, volumes, args...; kw...)
plot!(eos::BulkModulusEquation, volumes, args...; kw...)
plot!(plotobj, eos::EnergyEquation, volumes, args...; kw...)
plot!(plotobj, eos::PressureEquation, volumes, args...; kw...)
plot!(plotobj, eos::BulkModulusEquation, volumes, args...; kw...)

Plot the energy/pressure/bulk modulus versus volumes curve given an equation of state.

EquationOfStateRecipes.energyplot — Function

energyplot(params::Parameters, volumes, args...; kw...)
energyplot!(params::Parameters, volumes, args...; kw...)
energyplot!(plotobj, params::Parameters, volumes, args...; kw...)

Plot the energy versus volumes curves given the parameters of equations of state.

EquationOfStateRecipes.pressureplot — Function

pressureplot(params::Parameters, volumes, args...; kw...)
pressureplot!(params::Parameters, volumes, args...; kw...)
pressureplot!(plotobj, params::Parameters, volumes, args...; kw...)

Plot the pressure versus volumes curves given the parameters of equations of state.

EquationOfStateRecipes.bulkmodulusplot — Function

bulkmodulusplot(params::Parameters, volumes, args...; kw...)
bulkmodulusplot!(params::Parameters, volumes, args...; kw...)
bulkmodulusplot!(plotobj, params::Parameters, volumes, args...; kw...)

Plot the bulk modulus versus volumes curves given the parameters of equations of state.

EquationOfStateRecipes.energypressureplot — Function

energypressureplot(params::Parameters, volumes, args...; layout=(1, 2), kw...)
energypressureplot!(params::Parameters, volumes, args...; layout=(1, 2), kw...)
energypressureplot!(plotobj, params::Parameters, volumes, args...; layout=(1, 2), kw...)

Create a graph that shows the energy/pressure versus volume curves using the given parameters of equations of state on the same horizontal axis.

Warning

To use this function, you must specify the layout keyword argument as either (1, 2) or (2, 1)!