API Reference

For detailed examples, please refer to section "Examples".

EquationOfStateRecipes.energyplot — Function

energyplot(params::Parameters, volumes, args...; kw...)
energyplot!(params::Parameters, volumes, args...; kw...)
energyplot!(plotobj, params::Parameters, volumes, args...; kw...)

Plot the energy versus volumes curves given the parameters of equations of state.

EquationOfStateRecipes.pressureplot — Function

pressureplot(params::Parameters, volumes, args...; kw...)
pressureplot!(params::Parameters, volumes, args...; kw...)
pressureplot!(plotobj, params::Parameters, volumes, args...; kw...)

Plot the pressure versus volumes curves given the parameters of equations of state.

EquationOfStateRecipes.bulkmodulusplot — Function

bulkmodulusplot(params::Parameters, volumes, args...; kw...)
bulkmodulusplot!(params::Parameters, volumes, args...; kw...)
bulkmodulusplot!(plotobj, params::Parameters, volumes, args...; kw...)

Plot the bulk modulus versus volumes curves given the parameters of equations of state.

EquationOfStateRecipes.equationsplot — Function

equationsplot(params::Parameters, volumes, args...; kw...)
equationsplot!(params::Parameters, volumes, args...; kw...)
equationsplot!(plotobj, params::Parameters, volumes, args...; kw...)

Create a graph that shows the energy/pressure versus volume curves using the given parameters of equations of state on the same horizontal axis.